--- Sources/game/event/events/npc/starting_village/blacksmith/SVBlacksmithHelperNPC.swift
@@ -1,5 +1,0 @@
-enum SVBlacksmithHelperNPC: StartingVillageNPC {
-	static func talk() async {
-		await MessageBox.message("I'm busy here...", speaker: .blacksmith_helper)
-	}
-}
