--- Sources/game/event/events/npc/starting_village/builder/SVBuilderHelperNPC.swift
@@ -1,5 +1,0 @@
-enum SVBuilderHelperNPC: StartingVillageNPC {
-	static func talk() async {
-		await MessageBox.message("I'm busy here...", speaker: .builder_helper)
-	}
-}
