+++ Sources/game/event/events/npc/starting_village/carptenter/SVCarpenterHelperNPC.swift
@@ -1,5 +1,5 @@
 enum SVCarpenterHelperNPC: StartingVillageNPC {
 	static func talk() async {
-		await MessageBox.message("I'm busy here...", speaker: .carpenter_helper)
+		await MessageBox.message("I'm busy here...", speaker: .carpenter)
 	}
 }
