+++ Sources/game/event/events/npc/starting_village/carptenter/SVLeadCarpenterNPC.swift
@@ -1,14 +1,14 @@
-enum SVCarpenterNPC: StartingVillageNPC {
+enum SVLeadCarpenterNPC: StartingVillageNPC {
 	static func talk() async {
 		if await Game.shared.stages.blacksmith.stage3Stages == .goToCarpenter {
-			await MessageBox.message("Here are your sticks.", speaker: .carpenter)
+			await MessageBox.message("Here are your sticks.", speaker: .lead_carpenter)
 			if let lumberUUIDs = await Game.shared.stages.blacksmith.stage3LumberUUIDsToRemove {
 				await Game.shared.player.removeItems(ids: lumberUUIDs)
 			}
 			await Game.shared.stages.blacksmith.setStage3LumberUUIDsToRemove(Game.shared.player.collect(item: .init(type: .stick, canBeSold: false), count: 20))
 			await Game.shared.stages.blacksmith.setStage3Stages(.comeBack)
 		} else {
-			await MessageBox.message("I'm busy here...", speaker: .carpenter)
+			await MessageBox.message("I'm busy here...", speaker: .lead_carpenter)
 		}
 	}
 }
