--- Sources/game/event/events/npc/starting_village/farm/SVFarmerHelperNPC.swift
@@ -1,5 +1,0 @@
-enum SVFarmerHelperNPC: StartingVillageNPC {
-	static func talk() async {
-		await MessageBox.message("I'm busy here...", speaker: .farmer_helper)
-	}
-}
