--- Sources/game/event/events/npc/starting_village/mine/SVMineHelperNPC.swift
@@ -1,5 +1,0 @@
-enum SVMineHelperNPC: StartingVillageNPC {
-	static func talk() async {
-		await MessageBox.message("I'm busy here...", speaker: .mine_helper)
-	}
-}
