+++ Sources/models/game/buildings/protocols/NPCJobBuilding.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+protocol NPCJobBuilding: Equatable, Hashable, Identifiable, Sendable {
+	var workers: Set<UUID> { get }
+
+	func addWorker(_ worker: UUID) async
+	func removeWorker(_ worker: UUID) async
+}
